fix: sonar check and some adjusts
Extract "FileStream" in a new file.
Removed some includes and using forward declaration.
